[feat] splashactivity 추가
